+++ Scripts/Reports/Send-MaintenanceReport.ps1
@@ -37,14 +37,6 @@
     )
     begin {
         Write-Debug -Message ("[ Enter => function {0} ]" -f $MyInvocation.MyCommand);
-
-        Register-MSConfigStore;
-
-        Set-MSConfigDebugMode -Value ($false); # $true oder $false
-
-        Register-MSSession -UseDefaults ($true);
-        Select-MSSession -UseDefaults ($true);
-        Enter-MSSession -UseDefaults ($true);
     }
     process {
         Write-Debug -Message ("[ Process => function {0} ]" -f $MyInvocation.MyCommand);
@@ -75,10 +67,5 @@
     }
     end {
         Write-Debug -Message ("[ Leave => function {0} ]" -f $MyInvocation.MyCommand);
-
-        Exit-MSSession -UseDefaults ($true);
-        Unregister-MSSession -UseDefaults ($true);
-
-        Unregister-MSConfigStore;
     }
 }